pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock_prediction"
version = "0.1.0"
description = "Indian Stock Market Prediction System using ML and NLP"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    "torch>=2.2",
    "yfinance>=0.2.36",
    "feedparser>=6.0",
    "beautifulsoup4>=4.12",
    "requests>=2.31",
    "transformers>=4.38",
    "spacy>=3.7",
    "ollama>=0.4",
    "ta>=0.11",
    "click>=8.1",
    "rich>=13.7",
    "joblib>=1.3",
    "tqdm>=4.66",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff",
]

[project.scripts]
stockpredict = "stock_prediction.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
